refactor main file into separate class files
