+++ Race/animals.py
@@ -1,0 +1,19 @@
+class Rabbit:
+
+    speed = 10
+    weight = 1
+
+    # Initialize a new rabbit racer
+    def __init__(self, racername):
+        self.name = racername
+
+
+class Turtle:
+
+    speed = 2
+    weight = 5
+
+    # Initialize a new turtle racer
+    def __init__(self, racername):
+        self.name = racername
+
